feat: introduce the record package
The `record` package implements types and functions
to store Statuses in ringbuffers, per-node or per-process
(so possibly multiple nodes at once).

This is an additional optional package. Applications
using the fingerprint often needs to store statuses
to enable troubleshooting, so the `record` packages make
it easier to centralize this code.

There are no impacts on the rest of the codebase,
besides minor additions, again optional, to better
support the `record` flows, and some extra tests
added along the way.

Long story short, this new code is self-contained,
opt-in and as independent as possible.

Signed-off-by: Francesco Romani <fromani@redhat.com>